Moved http codes and request header utils to util\Http
