+++ src/response/Error.php
@@ -8,6 +8,7 @@ use alkemann\h2l\Log;
 use alkemann\h2l\Message;
 use alkemann\h2l\Request;
 use alkemann\h2l\Response;
+use alkemann\h2l\util\Http;
 
 /**
  * Class Error
@@ -60,7 +61,7 @@ class Error extends Response
         }
         $page_class = $this->config['page_class'];
 
-        $msg = Message::httpCodeToMessage($this->code);
+        $msg = Http::httpCodeToMessage($this->code);
         $h("HTTP/1.0 {$this->code} {$msg}");
         try {
             $page_config = $this->config + [
