+++ src/response/Json.php
@@ -3,6 +3,7 @@
 namespace alkemann\h2l\response;
 
 use alkemann\h2l\Message;
+use alkemann\h2l\util\Http;
 
 /**
  * Class Json
@@ -39,7 +40,7 @@ class Json extends \alkemann\h2l\Response
         $h = $this->config['header_func'] ?? 'header';
         $h("Content-type: application/json");
         if ($this->code != 200) {
-            $msg = Message::httpCodeToMessage($this->code);
+            $msg = Http::httpCodeToMessage($this->code);
             $h("HTTP/1.0 {$this->code} {$msg}");
         }
     }
