+++ src/util/ArrayManipulations.php
@@ -11,7 +11,6 @@ use OutOfBoundsException;
  */
 class ArrayManipulations
 {
-
     /**
      * Look for a deeo value in a nested data array.
      *
@@ -67,23 +66,5 @@ class ArrayManipulations
         } else {
             return self::getArrayValueByKeys($keys, $data[$key]);
         }
-    }
-
-    public static function getRequestHeadersFromServerArray(array $server_array)
-    {
-        $out = [];
-        foreach ($server_array as $name => $value) {
-            if (substr($name, 0, 5) == "HTTP_") {
-                $name = str_replace(" ", "-", ucwords(strtolower(str_replace("_", " ", substr($name, 5)))));
-                $out[$name] = $value;
-            }
-        }
-        if (array_key_exists("CONTENT_TYPE", $server_array)) {
-            $out["Content-Type"] = $server_array['CONTENT_TYPE'];
-        }
-        if (array_key_exists("CONTENT_LENGTH", $server_array)) {
-            $out["Content-Length"] = $server_array['CONTENT_LENGTH'];
-        }
-        return $out;
     }
 }
